fix: MSVC /W4 /WX clean build — standard C++20 portability fixes
Phase 5a of Item 20 (multi-platform build). 24 files, 295+/125-.

Bug fixes (all platforms):
- bitset.hpp toUlong(): floor division sizeof(unsigned long)/WORD_SIZE
  gave 0 on Windows LLP64 (ulong=4, word_t=8). Fixed with ceiling div.
  Silent on Linux LP64 where ulong=8.
- row_codec_delta002.hpp: out-of-bounds span access on empty string
  deserialization when len==0 at buffer boundary. Guarded with if(len>0).

Library (standard C++20, no pragmas/ifdefs):
- bitset.hpp, layout.hpp: else branches on if constexpr (C4702)
- row.hpp: arithmetic type guards on static_cast in get()/set() to
  prevent implicit int->string via operator=(char)
- vle.hpp: removed all 4 #pragma warning blocks, restructured with
  if constexpr else branches
- file_header.hpp: ternary replacing std::min for narrowing (C4244)
- lz4_stream.hpp: static_cast<int> for size_t->int
- bcsv_c_api.cpp: replaced __builtin_expect macro with [[unlikely]]
- bcsvValidate.cpp: tolower lambda for int->char narrowing

Build system:
- CMakePresets.json: ninja-msvc-debug/release presets
- CMakeLists.txt: conditional libm link (not on MSVC)
- tests/benchmark CMakeLists.txt: /bigobj for MSVC template depth

Tests:
- vectorized_access_test.cpp: portable volatile barrier (was GCC asm)
- visit_test.cpp, row_parameterized_test.cpp, bitset_test.cpp,
  bcsv_comprehensive_test.cpp: narrowing casts, constexpr guards,
  shadow renames, macro fixes
- bcsv_c_api_full_test.c: Windows portability (TMP_DIR, _mkdir, wchar)
- conftest.py: ninja-msvc preset discovery

Benchmarks:
- bench_common.hpp: NOMINMAX + windows.h guard
- bench_macro_datasets.cpp: removed 3 oversized LayoutStatic (132/200/
  290 cols) exceeding MSVC tuple recursion depth; dynamic mode remains

Verified: MSVC /W4 /WX 0 warnings, 694/694 GTest, 911/911 C API tests.
Needs regression check on GCC/Clang before merge.

+++ include/bcsv/lz4_stream.hpp
@@ -244,7 +244,7 @@ namespace bcsv {
                 int srcSize = static_cast<int>(src.size());
                 
                 while (true) {
-                    int dictLen = std::min(static_cast<size_t>(pos_), LZ4_DICT_SIZE);
+                    int dictLen = static_cast<int>(std::min(static_cast<size_t>(pos_), LZ4_DICT_SIZE));
                     int decompressedBytes = 0;
 
                     // Attempt 1: Append to Ring Buffer
